+++ v3/crates/plan-types/src/relationships.rs
@@ -3,6 +3,7 @@ use open_dds::{
     arguments::ArgumentName,
     commands::{CommandName, FunctionName},
     data_connector::DataConnectorColumnName,
+    models::ModelName,
     relationships::{RelationshipName, RelationshipType},
     types::{CustomTypeName, DataConnectorArgumentName},
 };
@@ -20,6 +21,7 @@ pub struct LocalModelRelationshipInfo<'s> {
     #[serde(serialize_with = "serialize_qualified_btreemap")]
     pub source_type_mappings:
         &'s BTreeMap<Qualified<CustomTypeName>, metadata_resolve::TypeMapping>,
+    pub target_model_name: &'s Qualified<ModelName>,
     pub target_source: &'s metadata_resolve::ModelSource,
     pub target_type: &'s Qualified<CustomTypeName>,
     pub mappings: &'s Vec<metadata_resolve::RelationshipModelMapping>,