+++ v3/crates/plan/src/order_by.rs
@@ -245,6 +245,7 @@ fn resolve_relationship_operand(
                 source_type: type_name,
                 source_data_connector: data_connector,
                 source_type_mappings: type_mappings,
+                target_model_name,
                 target_source: target_model_source,
                 target_type: &model_relationship_target.target_typename,
                 mappings: &model_relationship_target.mappings,